+++ app/src/main/java/name/monwf/customiuizer/MainFragment.java
@@ -7,7 +7,6 @@ import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
-import android.text.SpannableString;
 import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -31,8 +30,6 @@ import androidx.preference.PreferenceCategory;
 import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Locale;
 
 import name.monwf.customiuizer.prefs.ListPreferenceEx;
 import name.monwf.customiuizer.prefs.PreferenceEx;
@@ -42,6 +39,7 @@ import name.monwf.customiuizer.subs.Launcher;
 import name.monwf.customiuizer.subs.System;
 import name.monwf.customiuizer.subs.Various;
 import name.monwf.customiuizer.utils.AppHelper;
+import name.monwf.customiuizer.utils.AppLocaleController;
 import name.monwf.customiuizer.utils.Helpers;
 import name.monwf.customiuizer.utils.ModData;
 import name.monwf.customiuizer.utils.ModSearchAdapter;
@@ -264,44 +262,8 @@ public class MainFragment extends PreferenceFragmentBase {
 			}
 		});
 
-		String[] locales = new String[] { "zh-CN", "zh-TW", "ru-RU", "ja-JP", "vi-VN", "cs-CZ", "pt-BR", "tr-TR", "es-ES" };
-
-		ArrayList<String> localesArr = new ArrayList<String>(Arrays.asList(locales));
-		ArrayList<SpannableString> localeNames = new ArrayList<SpannableString>();
-		localesArr.add(0, "en");
-		for (String locale: localesArr) try {
-			Locale loc = Locale.forLanguageTag(locale);
-			StringBuilder locStr;
-			SpannableString locSpanString;
-			if (locale.equals("zh-TW")) {
-				locStr = new StringBuilder("繁體中文 (台灣)");
-			}
-			else {
-				locStr = new StringBuilder(loc.getDisplayLanguage(loc));
-				locStr.setCharAt(0, Character.toUpperCase(locStr.charAt(0)));
-				if (locale.equals("pt-BR")) {
-					locStr.append(" (Brasil)");
-				}
-			}
-			locSpanString = new SpannableString(locStr.toString());
-			localeNames.add(locSpanString);
-		} catch (Throwable t) {
-			localeNames.add(new SpannableString(Locale.getDefault().getDisplayLanguage(Locale.getDefault())));
-		}
-
-		localesArr.add(0, "auto");
-		localeNames.add(0, new SpannableString(getString(R.string.array_system_default)));
-
 		ListPreferenceEx locale = findPreference("pref_key_miuizer_locale");
-		locale.setEntries(localeNames.toArray(new CharSequence[0]));
-		locale.setEntryValues(localesArr.toArray(new CharSequence[0]));
-		locale.setOnPreferenceChangeListener(new CheckBoxPreference.OnPreferenceChangeListener() {
-			@Override
-			public boolean onPreferenceChange(Preference preference, Object newValue) {
-				getActivity().recreate();
-				return true;
-			}
-		});
+		AppLocaleController.setupLocalePreference(locale, AppHelper.appPrefs);
 
 	}